handles bigger list of cases

+++ pull_files.py
@@ -15,33 +15,38 @@ import pandas as pd
 import os
 
 df = pd.read_excel('../../other_data/list_of_cases.xlsx', parse_dates=[0])
-datetimes = [date.replace(hour=hour) for date, hour in zip(df.dates[df.selected == 'x'], df.h[df.selected == 'x'])]
+df.h.fillna(12, inplace=True)
+df.h = df.h.astype('int')
+datetimes = [date.replace(hour=hour) for date, hour in zip(df.dates, df.h)]
 ddir = r"C:\Users\sw825517\OneDrive - University of Reading\research\code\eumetsat"
 
 for start in datetimes:
-    dt = np.timedelta64(15, 'm')
+    if start.year == 2023:
+        dt = np.timedelta64(15, 'm')
 
-    # Retrieve datasets that match our filter
-    products = selected_collection.search(
-        dtstart=start,
-        dtend=start + dt)
+        # Retrieve datasets that match our filter
+        products = selected_collection.search(
+            dtstart=start,
+            dtend=start + dt)
 
-    for product in products:
-        print(str(product))
+        for product in products:
+            print(str(product))
 
-        if os.path.exists('data/' + start.strftime('%Y-%m-%d_%H')):
-            continue
-        else:
-            os.makedirs('data/' + start.strftime('%Y-%m-%d_%H'))
+            if os.path.exists('data/' + start.strftime('%Y-%m-%d_%H')):
+                continue
+            else:
+                os.makedirs('data/' + start.strftime('%Y-%m-%d_%H'))
 
-        with product.open() as fsrc, \
-                open(fsrc.name, mode='wb') as fdst:
-            shutil.copyfileobj(fsrc, fdst)
-            print(f'Download of product {product} finished.')
-            print(f'Unpacking {fdst.name}')
-            shutil.unpack_archive(fdst.name, extract_dir=ddir + '/data/' + start.strftime('%Y-%m-%d_%H'))
+            with product.open() as fsrc, \
+                    open(fsrc.name, mode='wb') as fdst:
+                shutil.copyfileobj(fsrc, fdst)
+                print(f'Download of product {product} finished.')
+                print(f'Unpacking {fdst.name}')
+                shutil.unpack_archive(fdst.name, extract_dir=ddir + '/data/' + start.strftime('%Y-%m-%d_%H'))
 
-        os.remove(fdst.name)
-        print(f'Removed {fdst.name}')
+            os.remove(fdst.name)
+            print(f'Removed {fdst.name}')
+    else:
+        pass
 
 print('All downloads are finished.')
